Use Aho-Corasick for Fuzzy Matching
Refactor op filtering to build a substring dictionary for each Op. These
dictionaries are then used to parse the search string. Each matched
fragment (substring) is worth 2*length - 1 points, and each Op's total
score is the sum of the score of its fragments. The longest possible
matches without overlap are emitted.

This is a much more robust mechanism for scoring as it enables arbitrary
fuzzy matches.