Create "Simple" view
Dumping info on all available Ops provides a lot of details that are
unnecessary and confusing for users. However, it is useful for
developers to be able to see what all Ops exist in a given installation.
Thus the Op Finder now sports both a Simple and Advanced view.

Advanced has full information for all Ops. In simple view we eliminate
non-image-related Ops and aggregate their display for common parameters.