Use icons for switching Modes
These images display both the words "Developer mode" and "User mode",
so that it is always clear what options are available, and which mode
you are currently in.

Previously the button would say "Dev mode" while in user mode, and vice
versa, as these were the modes you were switching to... now there is no
ambiguity.